Treat "403" status code as deployment gone
Adds proper handling of the "403" status code when the traffic filter
rule has already been deleted on on the update_traffic_rules file.

Also adds back the 403 state to be treated like the resource is gone.
