+++ ec/ecresource/trafficfilterassocresource/read.go
@@ -19,15 +19,13 @@ package trafficfilterassocresource
 
 import (
 	"context"
-	"errors"
-	"net/http"
 
 	"github.com/elastic/cloud-sdk-go/pkg/api"
-	"github.com/elastic/cloud-sdk-go/pkg/api/apierror"
 	"github.com/elastic/cloud-sdk-go/pkg/api/deploymentapi/trafficfilterapi"
-	"github.com/elastic/cloud-sdk-go/pkg/client/deployments_traffic_filter"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+
+	"github.com/elastic/terraform-provider-ec/ec/internal/util"
 )
 
 // read queries the remote deployment traffic filter ruleset association and
@@ -40,7 +38,7 @@ func read(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diag
 		IncludeAssociations: true,
 	})
 	if err != nil {
-		if ruleAssocNotFound(err) {
+		if util.TrafficFilterNotFound(err) {
 			d.SetId("")
 			return nil
 		}
@@ -52,16 +50,4 @@ func read(_ context.Context, d *schema.ResourceData, meta interface{}) diag.Diag
 	}
 
 	return nil
-}
-
-func ruleAssocNotFound(err error) bool {
-	// We're using the As() call since we do not care about the error value
-	// but do care about the error type since it's an implicit 404.
-	var ruleNotFound *deployments_traffic_filter.GetTrafficFilterRulesetNotFound
-	if errors.As(err, &ruleNotFound) {
-		return true
-	}
-
-	// We also check for the case where a 403 is thrown for ESS.
-	return apierror.IsRuntimeStatusCode(err, http.StatusForbidden)
 }
